requirements.txt
aiogram==3.2.0
asyncpg==0.28.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
requests==2.31.0
Pillow==10.1.0
pydantic==2.4.2
loguru==0.7.2
aiohttp==3.9.0